Separate the logic of selecting/deselecting from the fading animation that this selection has.
