+++ src/main/java/com/fundynamic/d2tm/game/entities/structures/Structure.java
@@ -1,7 +1,7 @@
 package com.fundynamic.d2tm.game.entities.structures;
 
 import com.fundynamic.d2tm.game.behaviors.Selectable;
-import com.fundynamic.d2tm.game.behaviors.SelectableImpl;
+import com.fundynamic.d2tm.game.behaviors.FadingSelection;
 import com.fundynamic.d2tm.game.entities.Entity;
 import com.fundynamic.d2tm.math.Vector2D;
 import org.newdawn.slick.Graphics;
@@ -11,7 +11,7 @@ import org.newdawn.slick.SpriteSheet;
 public class Structure extends Entity implements Selectable {
 
     // Behaviors
-    private SelectableImpl selectableImpl;
+    private FadingSelection fadingSelection;
 
     // Implementation
     private float animationTimer;
@@ -24,7 +24,7 @@ public class Structure extends Entity implements Selectable {
 
     public Structure(Vector2D mapCoordinates, SpriteSheet spriteSheet, int width, int height) {
         super(mapCoordinates, spriteSheet);
-        this.selectableImpl = new SelectableImpl(width, height);
+        this.fadingSelection = new FadingSelection(width, height);
     }
 
     public Image getSprite() {
@@ -38,7 +38,7 @@ public class Structure extends Entity implements Selectable {
         float offset = delta * ANIMATION_FRAMES_PER_SECOND;
         animationTimer = (animationTimer + offset) % ANIMATION_FRAME_COUNT;
 
-        this.selectableImpl.update(delta);
+        this.fadingSelection.update(delta);
     }
 
     public Vector2D getMapCoordinates() {
@@ -49,27 +49,27 @@ public class Structure extends Entity implements Selectable {
     public void render(Graphics graphics, int x, int y) {
         Image sprite = getSprite();
         graphics.drawImage(sprite, x, y);
-        selectableImpl.render(graphics, x, y);
+        fadingSelection.render(graphics, x, y);
     }
 
     public void select() {
-        selectableImpl.select();
+        fadingSelection.select();
     }
 
     public void deselect() {
-        selectableImpl.deselect();
+        fadingSelection.deselect();
     }
 
     @Override
     public boolean isSelected() {
-        return selectableImpl.isSelected();
+        return fadingSelection.isSelected();
     }
 
     @Override
     public String toString() {
         return "Structure{" +
                 "animationTimer=" + animationTimer +
-                ", selectable=" + selectableImpl +
+                ", selectable=" + fadingSelection +
                 '}';
     }
 }
