+++ src/main/java/com/fundynamic/d2tm/game/entities/units/Unit.java
@@ -1,7 +1,7 @@
 package com.fundynamic.d2tm.game.entities.units;
 
 import com.fundynamic.d2tm.game.behaviors.Selectable;
-import com.fundynamic.d2tm.game.behaviors.SelectableImpl;
+import com.fundynamic.d2tm.game.behaviors.FadingSelection;
 import com.fundynamic.d2tm.game.entities.Entity;
 import com.fundynamic.d2tm.math.Random;
 import com.fundynamic.d2tm.math.Vector2D;
@@ -10,25 +10,28 @@ import org.newdawn.slick.Image;
 import org.newdawn.slick.SpriteSheet;
 
 public class Unit extends Entity implements Selectable {
-    private final SelectableImpl selectable;
+    // Behaviors
+    private final FadingSelection fadingSelection;
+
+    // Implementation
     private final int facing;
 
     public Unit(Vector2D mapCoordinates, Image image, int width, int height) {
-        this(mapCoordinates, new SpriteSheet(image, width, height), new SelectableImpl(width, height));
+        this(mapCoordinates, new SpriteSheet(image, width, height), new FadingSelection(width, height));
     }
 
-    public Unit(Vector2D mapCoordinates, SpriteSheet spriteSheet, SelectableImpl selectable) {
+    public Unit(Vector2D mapCoordinates, SpriteSheet spriteSheet, FadingSelection fadingSelection) {
         super(mapCoordinates, spriteSheet);
         int possibleFacings = spriteSheet.getHorizontalCount();
         this.facing = Random.getRandomBetween(0, possibleFacings);
-        this.selectable = selectable;
+        this.fadingSelection = fadingSelection;
     }
 
     @Override
     public void render(Graphics graphics, int x, int y) {
         Image sprite = getSprite();
         graphics.drawImage(sprite, x, y);
-        this.selectable.render(graphics, x, y);
+        this.fadingSelection.render(graphics, x, y);
     }
 
 
@@ -45,21 +48,21 @@ public class Unit extends Entity implements Selectable {
 
     @Override
     public void select() {
-        selectable.select();
+        fadingSelection.select();
     }
 
     @Override
     public void deselect() {
-        selectable.deselect();
+        fadingSelection.deselect();
     }
 
     @Override
     public boolean isSelected() {
-        return selectable.isSelected();
+        return fadingSelection.isSelected();
     }
 
     @Override
     public void update(float deltaInMs) {
-        this.selectable.update(deltaInMs);
+        this.fadingSelection.update(deltaInMs);
     }
 }
